// src/infra/enum/verticalHeader-enum.ts
export enum VerticalHeader {
    VIEWS = 'Views',
    EARNING ='Earning',
    VIDEOS = 'Videos',
    FANS = 'Fans',
    NEW_FANS = 'New Fans',
    VIEWS_VIDEO = 'Views / video',
    AVG_WATCH_TIME = 'Avg. Watch Time',
    EARNING_VIDEO = 'Earning / video',
    ENGAGED_PEOPLE = 'Engaged people',
    SPENDINGS = 'Spendings',
    PROFITABILITY = 'Profitability'
}